Preliminary refactors
