+++ README.md
@@ -56,6 +56,8 @@ To run the tests outside docker, follow these steps:
    - `DEFENDANT_NAME`: The full name of the above defendant
    - `DEFENDANT_DOB`: The date of birth of the above defendant
 
+  Note that if you want to run the tests against the dockerised version of VCD you can point `VCD_URL` at `localhost:3001` (you will also need to ensure your local env vars mirror those in the docker-compose)
+
 4. Run the tests.
    Use the Playwright test runner UI:
    ```
