+++ config.js
@@ -7,6 +7,8 @@ dotenv.config({ path: envFile })
 export const VCD_URL = process.env.VCD_URL
 export const EMAIL = process.env.EMAIL
 export const PASSWORD = process.env.PASSWORD
+export const MANAGER_EMAIL = process.env.MANAGER_EMAIL
+export const MANAGER_PASSWORD = process.env.MANAGER_PASSWORD
 export const URN = process.env.URN
 export const ASN = process.env.ASN
 export const DEFENDANT_NAME = process.env.DEFENDANT_NAME
