+++ pages/search_page.js
