+++ tests/link_unlink_defendant.spec.js
@@ -1,10 +1,9 @@
 import { test, expect } from '@playwright/test'
-import { VCD_URL, EMAIL, PASSWORD, URN, DEFENDANT_NAME } from '../config.js'
-import { SigninPage } from './signin_page.js'
+import { URN, DEFENDANT_NAME } from '../config.js'
+import { SigninPage } from '../pages/signin_page.js'
+import { CaseSummaryPage } from '../pages/case_summary_page.js'
 
 const MAAT_ID = '6079985'
-let prosecutionCaseUrl = `${VCD_URL}/prosecution_cases/${URN}`
-
 test.describe.configure({ mode: 'serial' })
 
 test.describe('Link and Unlink defendant flow', () => {
@@ -16,8 +15,8 @@ test.describe('Link and Unlink defendant flow', () => {
     sharedContext = await browser.newContext()
     sharedPage = await sharedContext.newPage()
 
-    const signinPage = new SigninPage(sharedPage, VCD_URL)
-    await signinPage.signin(EMAIL, PASSWORD)
+    const signinPage = new SigninPage(sharedPage)
+    await signinPage.signInAsCaseworker()
   })
 
   // Clean up the shared context after all tests
@@ -28,19 +27,20 @@ test.describe('Link and Unlink defendant flow', () => {
   })
 
   test('step 1: checks "Not linked" status', async () => {
-    await sharedPage.goto(prosecutionCaseUrl)
+    await (new CaseSummaryPage(sharedPage)).goto(URN)
     await expect(sharedPage.locator('body')).toContainText('Not linked')
   })
 
   test.describe('step 2: Link a defendant to a Court Data', () => {
-    test('Link using an invalid MAAT ID', async () => {
-      await sharedPage.goto(prosecutionCaseUrl)
-
+    test ('Link page can be accessed', async() => {
+      await (new CaseSummaryPage(sharedPage)).goto(URN)
       await sharedPage.getByRole('link', { name: DEFENDANT_NAME })
                 .click();
 
       await expect(sharedPage).toHaveTitle(/^Defendant details/)
+    })
 
+    test('Link using an invalid MAAT ID', async () => {
       await sharedPage.getByLabel('MAAT ID')
                 .fill('123456')
 
@@ -59,25 +59,25 @@ test.describe('Link and Unlink defendant flow', () => {
       await expect(sharedPage.locator('body')).toContainText('You have successfully linked to the court data source')
       await expect(sharedPage.locator('body')).toContainText(MAAT_ID)
     })
+  })
 
-    test.describe('step 3: Unlink the defendant from Court Data', async () => {
-      test('Unlink the defendant', async () => {
-        await sharedPage.goto(prosecutionCaseUrl)
+  test.describe('step 3: Unlink the defendant from Court Data', async () => {
+    test('Unlink the defendant', async () => {
+      await (new CaseSummaryPage(sharedPage)).goto(URN)
 
-        await sharedPage.getByRole('link', { name: DEFENDANT_NAME })
-                  .click();
+      await sharedPage.getByRole('link', { name: DEFENDANT_NAME })
+                .click();
 
-        await sharedPage.locator('summary:has-text("Remove link to court data")') // Expand the summary section
-                  .click();
+      await sharedPage.locator('summary:has-text("Remove link to court data")') // Expand the summary section
+                .click();
 
-        await sharedPage.getByLabel('Reason for unlinking')
-                  .selectOption('1')
+      await sharedPage.getByLabel('Reason for unlinking')
+                .selectOption('1')
 
-        await sharedPage.getByRole('button', { name: 'Remove link to court data' })
-                  .click();
+      await sharedPage.getByRole('button', { name: 'Remove link to court data' })
+                .click();
 
-        await expect(sharedPage.locator('body')).toContainText('You have successfully unlinked from the court data source')
-      })
+      await expect(sharedPage.locator('body')).toContainText('You have successfully unlinked from the court data source')
     })
   })
 })
